feat: add SSE server
